Swap out generate_bookshelf.py for a Rust implementation
